+++ src/useFieldArray.ts
@@ -367,8 +367,9 @@ export function useFieldArray<
     });
 
     control._names.focus &&
-      focusFieldBy(control._fields, (key: string) =>
-        key.startsWith(control._names.focus),
+      focusFieldBy(
+        control._fields,
+        (key) => !!key && key.startsWith(control._names.focus),
       );
 
     control._names.focus = '';